[INFRA] Move `seqan3::search` from `search/algorithm/` to `search/`.
