+++ doc/tutorial/read_mapper/read_mapper_step2.cpp
@@ -8,7 +8,7 @@
 #include <seqan3/argument_parser/all.hpp>
 #include <seqan3/core/debug_stream.hpp>
 #include <seqan3/io/sequence_file/input.hpp>
-#include <seqan3/search/algorithm/all.hpp>
+#include <seqan3/search/all.hpp>
 
 //! [solution]
 struct reference_storage_t
