+++ doc/tutorial/read_mapper/read_mapper_step4.cpp
@@ -10,7 +10,7 @@
 #include <seqan3/argument_parser/all.hpp>
 #include <seqan3/io/alignment_file/output.hpp>
 #include <seqan3/io/sequence_file/input.hpp>
-#include <seqan3/search/algorithm/all.hpp>
+#include <seqan3/search/all.hpp>
 #include <seqan3/std/span>
 
 struct reference_storage_t
