+++ doc/tutorial/search/search_small_snippets.cpp
@@ -8,7 +8,7 @@ seqan3::cleanup index_file{"index.file"};
 #endif //SEQAN3_WITH_CEREAL
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/search/algorithm/search.hpp>
+#include <seqan3/search/search.hpp>
 #include <seqan3/search/fm_index/bi_fm_index.hpp>
 #include <seqan3/search/fm_index/fm_index.hpp>
 #include <seqan3/std/filesystem>
