+++ doc/tutorial/search/search_solution5.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alignment/configuration/all.hpp>
 #include <seqan3/alignment/pairwise/align_pairwise.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/search/algorithm/search.hpp>
+#include <seqan3/search/search.hpp>
 #include <seqan3/std/span>
 
 using seqan3::operator""_dna4;
