+++ include/seqan3/search/detail/search_common.hpp
