+++ include/seqan3/search/detail/search_scheme_precomputed.hpp
