+++ include/seqan3/search/detail/search_traits.hpp
