+++ include/seqan3/search/search_result_range.hpp
