+++ test/unit/search/search_configuration_test.cpp
@@ -7,7 +7,7 @@
 
 #include <type_traits>
 
-#include <seqan3/search/algorithm/all.hpp>
+#include <seqan3/search/all.hpp>
 
 #include <gtest/gtest.h>
