+++ test/unit/search/search_scheme_algorithm_test.cpp
@@ -12,8 +12,8 @@
 #include "helper_search_scheme.hpp"
 
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/search/algorithm/detail/search_scheme_algorithm.hpp>
-#include <seqan3/search/algorithm/detail/search_trivial.hpp>
+#include <seqan3/search/detail/search_scheme_algorithm.hpp>
+#include <seqan3/search/detail/search_trivial.hpp>
 #include <seqan3/search/fm_index/all.hpp>
 #include <seqan3/range/views/slice.hpp>
 #include <seqan3/range/views/to.hpp>
